Add treasure: cup_noodles, decrease 2 max hp and heal 6 hp

+++ tower-defense/src/game_state/mod.rs
@@ -165,7 +165,7 @@ impl GameState {
     }
 
     pub fn max_hp(&self) -> f32 {
-        self.config.player.max_hp + self.upgrade_state.max_hp_plus() as f32
+        self.config.player.max_hp + self.upgrade_state.max_hp_plus()
     }
 
     pub fn max_dice_chance(&self) -> usize {
